Allow for reusing TCP connections
This eliminates TCP handshakes, TCP slow start, and potentially TLS
negotiation for subsequent requests.

The `max_idle_pool_size` is just a starter value. I normally parse it
from the URL the way the crystal-lang/crystal-db shard does[1], but I
think it makes sense to discuss that as a feature first.

[1] https://github.com/crystal-lang/crystal-db/blob/eaddae7d71d52536453ea2d94777854f1b057f83/src/db/driver.cr#L31-L40